Create addToFavorite in favorites controller
