+++ server/controllers/recipes.js
@@ -1,7 +1,6 @@
 import models from '../models';
 
 const recipe = models.Recipe;
-const review = models.Review;
 
 /**
  * @exports createRecipe
@@ -168,51 +167,4 @@ export const deleteRecipe = (req, res) => {
     .catch(() => res.status(401).send({ error: 'Error Deleting Recipe' }));
 
   return deletedRecipe;
-};
-
-/**
- * @exports postReview
- * @param  {obj} req request object
- * @param  {obj} res result object
- * @return {obj}  newUser object
- */
-export const postReview = (req, res) => {
-  const userId = req.session.user.id;
-  const recipeId = req.params.recipeId;
-  const content = req.body.content;
-  const newReview = review
-    .create({
-      content,
-      userId,
-      recipeId
-    })
-    .then((createdReview) => {
-      res.status(201).send(createdReview);
-    })
-    .catch(() => res.status(401).send({ error: 'Error Posting Review' }));
-
-  return newReview;
-};
-
-/**
- * @exports getReviews
- * @param  {obj} req request object
- * @param  {obj} res result object
- * @return {obj}  newUser object
- */
-export const getReviews = (req, res) => {
-  const recipeId = req.params.recipeId;
-  const newReview = review
-    .findAll({
-      where: { recipeId },
-      include: [
-        { model: models.User, attributes: ['name'] }
-      ]
-    })
-    .then((reviews) => {
-      res.status(201).send(reviews);
-    })
-    .catch(() => res.status(401).send({ error: 'Error Fetching Reviews' }));
-
-  return newReview;
 };
