+++ server/routes/recipes/index.js
@@ -1,6 +1,7 @@
 import express from 'express';
 
 import * as recipeController from '../../controllers/recipes';
+import * as reviewController from '../../controllers/reviews';
 import * as auth from '../../controllers/auth';
 
 const user = express.Router();
@@ -13,7 +14,7 @@ user.get('/', recipeController.getAllRecipes);
 user.put('/:recipeId', recipeController.modifyRecipe);
 user.delete('/:recipeId', recipeController.deleteRecipe);
 
-user.post('/:recipeId/reviews', recipeController.postReview);
-user.get('/:recipeId/reviews', recipeController.getReviews);
+user.post('/:recipeId/reviews', reviewController.postReview);
+user.get('/:recipeId/reviews', reviewController.getReviews);
 
 export default user;
